Return after winter hibernation.. Fixed structure with uploading/downloading models from server;  Added new UI to newsfeed and replies view; Deleted old classes;

+++ app/src/main/java/com/mateyinc/marko/matey/activity/Util.java
@@ -104,8 +104,8 @@ public class Util {
 //        return oS.toString();
 //    }
 
-    public static int parseDp(int value, Resources res) {
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, value, res.getDisplayMetrics());
+    public static float parseDp(float value, Resources res) {
+        return  TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, value, res.getDisplayMetrics());
     }
 
     /**
